scripts/requirements.txt
absl-py==0.12.0
astor==0.8.1
astroid==2.5.6
autopep8==1.5.7
boto3==1.17.96
botocore==1.20.96
cached-property==1.5.2
cachetools==4.2.2
certifi==2021.5.30
chardet==4.0.0
click==8.0.1
dnspython==1.16.0
Flask==2.0.1
future==0.18.2
gast==0.2.2
google-auth==1.31.0
google-auth-oauthlib==0.4.4
google-pasta==0.2.0
grpcio==1.38.0
h5py==3.2.1
idna==2.10
importlib-metadata==4.5.0
iso8601==0.1.14
isort==5.8.0
itsdangerous==2.0.1
Jinja2==3.0.1
jmespath==0.10.0
joblib==1.0.1
Keras==2.4.3
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
lazy-object-proxy==1.6.0
Markdown==3.3.4
MarkupSafe==2.0.1
mccabe==0.6.1
mock==4.0.3
numpy==1.20.3
oauthlib==3.1.1
opencv-python==4.5.1.48
opt-einsum==3.3.0
protobuf==3.17.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.7.0
pylint==2.8.3
pymongo==3.11.4
pyserial==3.5
python-dateutil==2.8.1
python-dotenv==0.17.1
PyYAML==5.4.1
requests==2.25.1
requests-oauthlib==1.3.0
RPi.GPIO==0.7.0
rsa==4.7.2
s3transfer==0.4.2
scikit-learn==0.24.2
scipy==1.6.3
serial==0.0.97
six==1.16.0
sklearn==0.0
tensorboard==2.0.2
tensorflow @ file:///home/pi/scripts/tensorflow-2.0.0-cp37-none-linux_armv7l.whl
tensorflow-estimator==2.0.1
termcolor==1.1.0
threadpoolctl==2.1.0
toml==0.10.2
typed-ast==1.4.3
typing-extensions==3.10.0.0
urllib3==1.26.5
Werkzeug==2.0.1
wrapt==1.12.1
zipp==3.4.1
